Install-Ninja: auto-detect hash type

+++ C++/Install-Ninja.psm1
@@ -8,24 +8,24 @@ Set-StrictMode -Version Latest
   Manages the installation of the Ninja build system.
   By version tag as published to GitHub.
 .EXAMPLE
-  Install-Ninja v1.8.2 -SHA512 9B9CE248240665FCD6404B989F3B3C27ED9682838225E6DC9B67B551774F251E4FF8A207504F941E7C811E7A8BE1945E7BCB94472A335EF15E23A0200A32E6D5
+  Install-Ninja v1.8.2 -Hash 9B9CE248240665FCD6404B989F3B3C27ED9682838225E6DC9B67B551774F251E4FF8A207504F941E7C811E7A8BE1945E7BCB94472A335EF15E23A0200A32E6D5
   -- Install Ninja-build v1.8.2 ...
   -- Install Ninja-build v1.8.2 ... done
 
   Download Ninja-build and install to `.\ninja-v1.8.2\`.
 .EXAMPLE
-  Install-Ninja -Tag v1.8.2 -SHA512 $SHA_hash -AddToPath
+  Install-Ninja -Tag v1.8.2 -HASH $SHA_hash -AddToPath
   -- Install Ninja-build v1.8.2 ...
   -- Install Ninja-build v1.8.2 ... done
 
   [Preferred usage]
   Add the ninja to the front of `$env:PATH`. To allow execution with: `ninja`.
 .EXAMPLE
-  $NinjaExe = Install-Ninja -Tag v1.8.2 -SHA512 $SHA_hash -AddToPath
+  $NinjaExe = Install-Ninja -Tag v1.8.2 -Hash $SHA_hash -AddToPath
 
   Save the path to `ninja.exe` in a variable.
 .EXAMPLE
-  Install-Ninja v1.8.2 -SHA512 $SHA_hash -InstallDir ~\tools\ninja
+  Install-Ninja v1.8.2 -Hash $SHA_hash -InstallDir ~\tools\ninja
   -- Install Ninja-build v1.8.2 ...
   -- Install Ninja-build v1.8.2 ... done
 
@@ -40,7 +40,7 @@ Set-StrictMode -Version Latest
   Install-Ninja -Tag v1.8.2 -Quiet
   WARNING: Install-Ninja: No hash given.
 .EXAMPLE
-  Install-Ninja -Tag v1.8.2 -SHA512 aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
+  Install-Ninja -Tag v1.8.2 -Hash aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
   -- Install Ninja-build v1.8.2 ...
   ERROR: Install-Ninja: download hash changed!
   Expected: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
@@ -51,22 +51,15 @@ function Install-Ninja {
   [OutputType([System.Management.Automation.PathInfo])]
 #  [OutputType([String])]
   param(
-    [Parameter(ParameterSetName='SHA256')]
-    [Parameter(ParameterSetName='SHA512')]
-    [Parameter(ParameterSetName='default')]
     [ValidateNotNullOrEmpty()]
+    [Alias('Version')]
     # GitHub release tag. E.g. "v1.9.0".
     $Tag = $(throw 'Tag is a required parameter'),
-    [Parameter(ParameterSetName='SHA512',Mandatory)]
-    [ValidateLength(128,128)]
+    [ValidatePattern('[0-9a-zA-Z]+')]
     [ValidateNotNullOrEmpty()]
-    # Sha512 hash for the downloaded archive.
-    [String]$SHA512,
-    [Parameter(ParameterSetName='SHA256',Mandatory)]
-    [ValidateLength(64,64)]
-    [ValidateNotNullOrEmpty()]
-    # SHA256 hash for the downloaded archive.
-    [String]$SHA256,
+    [Alias('SHA256', 'SHA512')]
+    # Hash matching the downloaded archive.
+    [String]$Hash,
     [ValidateScript({ Test-Path -LiteralPath "$_" -PathType Container })]
     [ValidateNotNullOrEmpty()]
     # An existing directory to install to (defaults to the current directory).
@@ -84,15 +77,26 @@ function Install-Ninja {
     Get-CommonFlagsCaller $PSCmdlet $ExecutionContext.SessionState
 
     if (-not $Quiet) { Write-Host "-- Install Ninja-build ${Tag} ..." }
-    if ($SHA512) {
-      $Hash_Type = 'SHA512'
-      $input_hash = $SHA512
-    } elseif ($SHA256) {
-      $Hash_Type = 'SHA256'
-      $input_hash = $SHA256
+
+    if ($Hash) {
+      switch ($Hash.Length) {
+        128 { $HashType = 'SHA512'; break }
+        64  { $HashType = 'SHA256'; break }
+        32  { $HashType = 'MD5'; break }
+        96  { $HashType = 'SHA384'; break }
+        40  { $HashType = 'SHA1'; break }
+        default {
+          Send-Message -Error (
+            $($MyInvocation.MyCommand).ToString() + ': Unsupported hash type!') `
+            -ContinueOnError:$Force -HideDetails:$Quiet `
+            -Details ('Length: ' + $Hash.Length),
+             'Supported hash types: SHA(1|256|384|512)|MD5'
+        }
+      }
     } else {
-      Send-Message -Warning "$($MyInvocation.MyCommand): No hash given."
-      $input_hash = $null
+      Send-Message -Warning (
+        $($MyInvocation.MyCommand).ToString() + ': No hash given.'
+      )
     }
     ($Path, $Temporary) = Join-Path ($InstallDir, $env:TEMP) ('ninja-' + $Tag)
     $FileName = 'ninja-win.zip'
@@ -103,14 +107,19 @@ function Install-Ninja {
   Process
   {
     if (Test-Path (Join-Path $Path 'ninja.exe')) {
-      Write-Verbose `
-        'Skip download and extraction. Already present in install location.'
-      return $null
+      if ($Force) {
+        Join-Path $Path 'ninja.exe' |
+          Send-Message -Warning 'Overwriting existing files'
+      } else {
+        Write-Verbose `
+          'Skip download and extraction. Already present in install location.'
+        return $null
+      }
     }
 
     if (-not (Test-Path $Archive) -or
-      ( $input_hash -and
-        ( $input_hash -ne (Get-FileHash $Archive -Algorithm $Hash_Type).Hash )
+      ( $Hash -and
+        ( $Hash -ne (Get-FileHash $Archive -Algorithm $HashType).Hash )
       )
     ) {
       $Err = $null
@@ -123,14 +132,14 @@ function Install-Ninja {
         )
       }
 
-      if ($input_hash -and -not $WhatIfPreference ) {
-        $hash = (Get-FileHash $Archive -Algorithm $Hash_Type).Hash
-        if ($input_hash -ne $hash) {
-          if (-not $Force) { Remove-Temporary $Temporary }
+      if ($Hash -and -not $WhatIfPreference ) {
+        $NewHash = (Get-FileHash $Archive -Algorithm $HashType).Hash
+        if ($Hash -ne $NewHash) {
+          Remove-Temporary $Temporary
           Send-Message -Error (
             ($MyInvocation.MyCommand).ToString() + ': download hash changed!') `
             -ContinueOnError:$Force -HideDetails:$Quiet `
-            -Details ('Expected: ' + $input_hash), ('Actual:   ' + $hash)
+            -Details ('Expected: ' + $Hash), ('Actual:   ' + $NewHash)
         }
       }
     }
@@ -159,17 +168,17 @@ function Install-Ninja {
       if (-not (Test-Command 'ninja --version') ) {
         '-- $env:Path --', ($env:Path -replace ';',"`n") |
           Send-Message -Error 'Failed to find Ninja on the search path.' `
-            -HideDetails:$Quiet
+            -HideDetails:$Quiet -ContinueOnError:$Force
       }
     }
 
     $Path = Join-Path $Path 'ninja.exe'
-    if (Test-Path $Path) {
-      return (Resolve-Path -LiteralPath $Path)
-    } else { return $null }
     if (-not $Quiet) {
       Write-Host "-- Install Ninja-build ${Tag} ... done"
     }
+    if (Test-Path $Path) {
+      return (Resolve-Path -LiteralPath $Path)
+    } else { return $null }
   }
 } #/ function Install-Ninja
 ##====--------------------------------------------------------------------====##
